MantineProvider only in root layout
