+++ apps/web/src/app/(authenticated)/layout.tsx
@@ -1,33 +1,11 @@
-import '@mantine/core/styles.css';
-import '@mantine/charts/styles.css';
-import { ColorSchemeScript, MantineProvider } from '@mantine/core';
-import type { Metadata } from 'next';
-import { NextIntlClientProvider } from 'next-intl';
-import { getLocale, getMessages } from 'next-intl/server';
 import { Suspense } from 'react';
-import { MainAppShell } from '@/components/shells/main-shell/main-shell';
 import LoaderCentered from '@/components/misc/loader-centered';
+import { MainAppShell } from '@/components/shells/main-shell/main-shell';
 
-export const metadata: Metadata = {
-  title: 'AdsViewer',
-  description: 'Get granular insights into your ads performance',
-};
-
-export default async function RootLayout({ children }: { children: React.ReactNode }): Promise<React.ReactNode> {
-  const locale = await getLocale();
-  const messages = await getMessages();
+export default function RootLayout({ children }: { children: React.ReactNode }): React.ReactNode {
   return (
-    <html lang={locale}>
-      <body>
-        <ColorSchemeScript />
-        <NextIntlClientProvider messages={messages}>
-          <MantineProvider>
-            <MainAppShell>
-              <Suspense fallback={<LoaderCentered />}>{children}</Suspense>
-            </MainAppShell>
-          </MantineProvider>
-        </NextIntlClientProvider>
-      </body>
-    </html>
+    <MainAppShell>
+      <Suspense fallback={<LoaderCentered />}>{children}</Suspense>
+    </MainAppShell>
   );
 }
