+++ apps/web/src/app/layout.tsx
@@ -1,14 +1,27 @@
-import './globals.scss';
+import '@mantine/core/styles.css';
+import '@mantine/charts/styles.css';
+import { ColorSchemeScript, MantineProvider } from '@mantine/core';
+import type { Metadata } from 'next';
+import { NextIntlClientProvider } from 'next-intl';
+import { getLocale, getMessages } from 'next-intl/server';
 
-export const metadata = {
-  title: 'Main',
-  description: 'I have followed setup instructions carefully',
+export const metadata: Metadata = {
+  title: 'AdsViewer',
+  description: 'Get granular insights into your ads performance',
 };
 
-export default function RootLayout({ children }: { children: React.ReactNode }): React.ReactNode {
+export default async function RootLayout({ children }: { children: React.ReactNode }): Promise<React.ReactNode> {
+  const locale = await getLocale();
+  const messages = await getMessages();
+
   return (
-    <html lang="en-gb">
-      <body>{children}</body>
+    <html lang={locale}>
+      <body>
+        <ColorSchemeScript />
+        <NextIntlClientProvider messages={messages}>
+          <MantineProvider>{children}</MantineProvider>
+        </NextIntlClientProvider>
+      </body>
     </html>
   );
 }
